requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
greenlet==1.1.2
numpy==1.21.3
pandas==1.3.4
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-telegram-bot==13.7
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.26
tornado==6.1
tzdata==2021.5
tzlocal==4.1
